rudimentary sentiment/possible n-gram
https://pypi.python.org/pypi/textblob

basically I think the sentiment works but it doesn’t score very well
and there is this error where ‘c-4’ breaks the sentiment system and i
tried to implement a cleaner solution but it was taking forever and so
i gave up for the night but thought it might be good just to push this
for now since it doesn’t break anything but maybe one of you guys can
come up with a quick fix for my stuff/check it out to make sure it’s
not wrong.

also took a shot at n-gram but once again might have been thwarted… but
since i didn’t break what’s currently there i decided to push this too
so someone can take a look at it. we can always revert this whole
commit if it is all shit